+++ src/styling/utils.ts
@@ -1,4 +1,4 @@
-import cssColorNames from './css-color-names';
+import cssColorNames from './css-color-names'
 import type { CssDictionary } from './css-dictionary'
 
 export function getStyleString(styles: CssDictionary) {
@@ -14,9 +14,11 @@ export function colorToHex(color: string | undefined): string | undefined {
 }
 
 export function colorToRgb(
-  color: string | undefined
+  color: string | string[] | undefined
 ): [number, number, number] | undefined {
-  if (!color || color === 'transparent') return undefined
+  // We ignore tiling patterns (arrays)
+  if (!color || Array.isArray(color) || color === 'transparent')
+    return undefined
   const hexColor = colorToHex(color)
   if (!hexColor) return undefined
   return [
@@ -27,14 +29,15 @@ export function colorToRgb(
 }
 
 export function opacityColor(
-  color: string | null | undefined,
+  color: string | string[] | null | undefined,
   opacity: number | null | undefined
 ): string | undefined {
-  if (color == null) return undefined
+  // We ignore tiling patterns (arrays)
+  if (color == null || Array.isArray(color)) return undefined
   if (opacity == null || opacity == 1) return color
   if (opacity == 0) return 'transparent'
   const rgb = colorToRgb(color)
-  if (!rgb) return color;
+  if (!rgb) return color
 
   return `rgba(${rgb.join(', ')}, ${opacity})`
 }